web/requirements.txt
Flask==2.1.2
Flask_Cors==3.0.10
Flask_Docs==0.7.2
Flask_SQLAlchemy==2.5.1
gevent==21.12.0
PyMySQL==1.0.2
python_dateutil==2.8.2
simplejson==3.17.6
SQLAlchemy==1.4.38
